Removed the remaining components of local encryption
For sessions.
