+++ config/autoload/global.php
@@ -101,16 +101,6 @@ return [
             'auto_create' => getenv('OPG_LPA_COMMON_DYNAMODB_AUTO_CREATE') ?: false,
         ],
 
-        'encryption' => [
-            'enabled' => true,
-            // Keys must be in the format: <ident: int> => <key: 32 character ASCII string>
-            'keys' => getenv('OPG_LPA_FRONT_SESSION_ENCRYPTION_KEYS') ?
-                array_combine(
-                    array_map( function( $v ){ return explode(':', trim($v))[0]; } , explode(',', trim(getenv('OPG_LPA_FRONT_SESSION_ENCRYPTION_KEYS')))),
-                    array_map( function( $v ){ return explode(':', trim($v))[1]; } , explode(',', trim(getenv('OPG_LPA_FRONT_SESSION_ENCRYPTION_KEYS'))))
-                ) : array(),
-        ],
-
     ], // session
 
     'csrf' => [
